+++ packages/fabric/platform.d.ts
@@ -1,4 +1,5 @@
 declare const platform: {
+  PREBUILT_HOSTS: string[]
   platformPackageName(platform?: string, arch?: string): string | null
   resolvePlatformPrebuilds(): string | null
   runtimeHost(): { platform: string | undefined, arch: string | undefined }